Create new layout. Provide widget ids.
On Tues. May 28, the topic of how to implement launch logic into
ltc.py was discussed.  It was decided that the main app window be
divided into three section:
  launch state info
  power and launch controls
  sensor indicators

ltc.py now has three primary divisions in the main window. Each
has an id defined in ltc.kv.  An example of using that id is
found at the end of the build method of LTCApp.