WIP update of LTC dependencies
It's been a decade since any significant work was done and there's been
quite a bit of changes in the software landscape. The biggest one is
python 2 is gone, but the other two major dependencies (Kivy, Phidget)
have backwards incompatible versions as well. This commit is the first
step at modernizing it:
- Convert to Python 3. Relatively straight forward, we didn't hit any
  backwards incompatible bits beyond print() so it is mostly just
  updating to a modern style
- Update Kivy to 2.3.1. There's been a lot of things deprecated and
  removed from Kivy but it looks like it works with the subset of
  widgets we have. There's a deprecation warning about templates in the
  kv file but those can be fixed later
- Update to Phidget22. Completely changed how the library works and a
  lot of rewriting going into this
- Added a pyproject.toml

I've not actually tested this on real Phidget hardware yet so no idea if
it works but launching the GUI doesn't crash and I can click the
buttons. Committing this now to have others work on this with me.