debug errors
IOIndicator tries to set val=event.value, but raises an exception every
time. It then tries to set val=event.state, but event is an instance of
a VoltageSensor, not really an "event". Setting val = 0 makes all the
errors go away, but surely this is not what was intended. More work
needs to be done. The app seems to struggle with opening the Relay
channels. Making the ignition a DigitalOutput and setting its devserial
and channel will allow it to open. We'll need to figure out what part of
Relay or LTCPhidget classes prevent it from opening.