Remove calls to setDeviceLabel, delete debug file
I didn't mean to check ltc_r.py into git, that file was just for debug
purposes. setDeviceLabel is actually used to set the device to open a
device with a certain label, not to write the label to flash. Using this
method is the reason that openWaitForAttachment was failing.