Continue bugfixes
There's something weird about the on_value hanlders where it was getting
the reference to self and passing that to the callbacks rather than the
actual value of the sensors. The args we were calling "unknown" were the
actual sensor values. The sensors were also not opening properly because
they were not assigning their channels, so I fixed that. Now they open,
but the values aren't being displayed properly. That will be my next
step.