Use builtin conversion functions
in phidget22 we can just set the type of sensor and use getSensorValue()
to convert rather than using our own function to convert. The values
seem more reasonable but still probably not accurate, so more work needs
to be done. Also, because this is so much simpler we can probably
refactor a lot of the code involved with displaying these sensor values.