Adjust error handling to allow clear messages in StatusDisplay
Before, StatusDisplay would just say that a phidget call failed. This
was happening on all errors, even normal ones like a sensor value being
out of normal range. This made the status display only ever say the same
"phidget call failed" message, which was very unhelpful. To improve
error handling I've created an LTCError class that extends Exception,
which linters seem to like a lot better. I've also adjusted the way
errors get sent to StatusDisplay so that we can if desired add a message
to be displayed. There was also a seg fault happening when we tried to
access "self.states[state][1]" when the state arg is a string.