Refactor relay nominal_value method
Since we don't get strings anymore in nominal_value, we don't need to
save nominal string values. We can simply save whether the relay is
nominal open or nominal closed, and then if the bool we get in
nominal_value equals that member variable then the relay is not nominal.