Ensure IOIndicators are intitialized with correct states
When the program was starting many of the ioindicators displayed
incorrect or misleading values. This commit makes sure that the
indicators properly convey their readings in the GUI. This is mainly
achieved by rewriting the set_state method on RelayLabel, and making
sure the relays are closed on startup. They should always be closed
anyway because they get shut off on exit, but this makes sure the
ioindicators are aware of the starting state of the relays.